chore: move carousel as comp
